Add some typecasts for intptr_t and uintptr_t value to fix gcc warnings.

+++ source/blender/src/sequence.c
@@ -2959,7 +2959,7 @@ void do_render_seq(RenderResult *rr, int cfra)
 
 			if (max != 0 && mem_in_use + mmap_in_use > max) {
 				fprintf(stderr, "mem_in_use = %lu, max = %lu\n",
-					mem_in_use + mmap_in_use, max);
+					(long unsigned int)mem_in_use + mmap_in_use, (long unsigned int)max);
 				fprintf(stderr, "Cleaning up, please wait...\n"
 					"If this happens very often,\n"
 					"consider "
